Style links in the calendar the same as everywhere else
This reverts 7d7d19472 (Make links in event descriptions the right
colour, 2021-03-05), which achieved less of the goal in a less
maintainable way. Instead of overriding the overridden style in order to
reapply the global styling, we now just more narrowly scope the override
for the event times and titles.